game-user-relation and global AuthGuard

+++ src/games/games.service.ts
@@ -1,5 +1,6 @@
-import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
+import { ConflictException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { plainToInstance } from 'class-transformer';
 import { Repository } from 'typeorm';
 import { CreateGameDto } from './dto/create-game.dto';
 import { UpdateGameDto } from './dto/update-game.dto';
@@ -9,38 +10,40 @@ import { Game } from './entities/game.entity';
 export class GamesService {
   constructor(@InjectRepository(Game) private readonly gamesRepository: Repository<Game>) { }
 
-  async create(createGameDto: CreateGameDto): Promise<Game> {
+  async create(authorId: string, createGameDto: CreateGameDto): Promise<Game> {
     if (await this.gamesRepository.existsBy({ title: createGameDto.title })) throw new ConflictException('Game with this title already exists!');
 
-    return await this.gamesRepository.save(this.gamesRepository.create(createGameDto));
+    return await this.gamesRepository.save(this.gamesRepository.create({ ...createGameDto, author: { id: authorId } }));
   }
 
   async findAll(): Promise<Game[]> {
-    return await this.gamesRepository.find();
+    return await this.gamesRepository.find({ relations: { author: true } });
   }
 
   async findOne(id: string): Promise<Game> {
-    const game: Game | null = await this.gamesRepository.findOne({ where: { id: id } });
+    const game: Game | null = await this.gamesRepository.findOne({ where: { id: id }, relations: { author: true } });
 
     if (!game) throw new NotFoundException();
 
     return game;
   }
 
-  async update(id: string, updateGameDto: UpdateGameDto): Promise<Game> {
+  async update(authorId: string, id: string, updateGameDto: UpdateGameDto): Promise<Game> {
     if (await this.gamesRepository.existsBy({ title: updateGameDto.title })) throw new ConflictException('Game with this title already exists!');
 
-    const game: Game | null = await this.gamesRepository.findOne({ where: { id: id } });
+    const game: Game | null = await this.gamesRepository.findOne({ where: { id: id }, relations: { author: true } });
 
     if (!game) throw new NotFoundException();
+    if (game.author.id !== authorId) throw new ForbiddenException();
 
-    return await this.gamesRepository.save({ ...game, ...updateGameDto });
+    return plainToInstance(Game, await this.gamesRepository.save({ ...game, ...updateGameDto }));
   }
 
-  async remove(id: string): Promise<Game> {
-    const game: Game | null = await this.gamesRepository.findOne({ where: { id: id } });
+  async remove(authorId: string, id: string): Promise<Game> {
+    const game: Game | null = await this.gamesRepository.findOne({ where: { id: id }, relations: { author: true } });
 
     if (!game) throw new NotFoundException();
+    if (game.author.id !== authorId) throw new ForbiddenException();
 
     return await this.gamesRepository.remove(game);
   }
